+++ tests/compilation_failures/tokens/intersperse_bool_input_braces_issue.rs
@@ -1,0 +1,13 @@
+use preinterpret::*;
+
+fn main() {
+    preinterpret! {
+        [!intersperse! {
+            items: [1 2],
+            separator: [],
+            add_trailing: {
+                true false
+            }
+        }]
+    }
+}