+++ tests/compilation_failures/tokens/intersperse_stream_input_variable_issue.rs
@@ -1,0 +1,11 @@
+use preinterpret::*;
+
+fn main() {
+    preinterpret! {
+        [!set! #x = 1 2]
+        [!intersperse! {
+            items: #..x,
+            separator: []
+        }]
+    }
+}